requirements.txt
boto3==1.28.48
botocore==1.31.48
fastapi==0.103.1
openai==0.28.0
python-dotenv==1.0.0
uvicorn==0.23.2
